Altering project structure
