+++ output/data/results.csv
@@ -1,6 +1,6 @@
 project;merge commit;class;method;left modifications;left deletions;right modifications;right deletions
 testRepository;de96a282565510196a5b0d81fc92e91c1e909428;Test;multipleCommits();[34, 35];[];[32];[]
-testRepository;1598887a045a86222b6d96dd1357b6a08f232ffa;Test;deletionBothAndAdditionBoth();[17];[];[20];[]
 testRepository;1598887a045a86222b6d96dd1357b6a08f232ffa;Test;deletionAndAdditionBoth();[25];[];[27];[]
-testRepository;10fe406824dd71ab9e14a3418d7bda4ab53e32a1;Test;deletionSingleAndAddition();[];[5];[6];[]
+testRepository;1598887a045a86222b6d96dd1357b6a08f232ffa;Test;deletionBothAndAdditionBoth();[17];[];[20];[]
 testRepository;10fe406824dd71ab9e14a3418d7bda4ab53e32a1;Test;deletionBothAndAddition();[];[11];[12];[11]
+testRepository;10fe406824dd71ab9e14a3418d7bda4ab53e32a1;Test;deletionSingleAndAddition();[];[5];[6];[]
