+++ src/main/arguments/ArgsParser.groovy
@@ -1,7 +1,10 @@
+package main.arguments
+
 import groovy.cli.commons.CliBuilder
 import groovy.cli.commons.OptionAccessor
 import java.text.SimpleDateFormat
 import java.text.ParseException
+import main.exception.InvalidArgsException
 
 class ArgsParser {
