+++ src/main/exception/InvalidArgsException.groovy
@@ -1,3 +1,5 @@
+package main.exception
+
 public class InvalidArgsException extends Exception {
     public InvalidArgsException(String message) {
         super(message)
