+++ src/main/project/MergeCommit.groovy
@@ -1,3 +1,5 @@
+package main.project
+
 class MergeCommit {
 
     private String sha
