+++ src/main/util/FileManager.groovy
@@ -1,9 +1,13 @@
+package main.util
+
 @Grab(group='commons-io', module='commons-io', version='2.6')
 import java.nio.file.Files 
 import java.nio.file.Paths
 import java.nio.file.Path
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import org.apache.commons.io.FileUtils 
+
+import main.project.*
 
 final class FileManager {
