+++ src/main/util/ProcessRunner.groovy
@@ -1,3 +1,5 @@
+package main.util
+
 final class ProcessRunner {
 
     public static Process runProcess(String directory, String... command) {
